% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultility.R
\name{ARI}
\alias{ARI}
\title{Compare two clustering with the Adjusted Rand Index}
\usage{
ARI(x, y)
}
\arguments{
\item{x}{A vector of integers, the clusters labels}

\item{y}{A vector of integers of the same length as x, the clusters labels}
}
\value{
A number between 0 (random clustering) and 1 (identical clustering)
}
\description{
Compare two clustering with the Adjusted Rand Index
}
\examples{
ARI(x = c(1, 2, 1), y = c(2, 2, 1))
}
