% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{build_fold_matrix}
\alias{build_fold_matrix}
\title{Title}
\usage{
build_fold_matrix(X, K)
}
\arguments{
\item{X}{An adjacency matrix}

\item{K}{An integer, the number of folds}
}
\value{
A matrix of the same size than X with class integer as coefficient
}
\description{
Title
}
