% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGenMLVSBM-class.R
\name{coef.FitGenMLVSBM}
\alias{coef.FitGenMLVSBM}
\title{Extract model coefficients}
\usage{
\method{coef}{FitGenMLVSBM}(object, ...)
}
\arguments{
\item{object}{an R6 object of class FitMLVSBM}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
List of parameters.
}
\description{
Extracts model coefficients from objects with class
\code{\link[=FitGenMLVSBM]{FitGenMLVSBM}}
}
