% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGenMLVSBM-class.R, R/FitMLVSBM-class.R
\name{coef.FitMLVSBM}
\alias{coef.FitMLVSBM}
\title{Extract model coefficients}
\usage{
\method{coef}{FitMLVSBM}(object, ...)

\method{coef}{FitMLVSBM}(object, ...)
}
\arguments{
\item{object}{an R6 object of class FitMLVSBM}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
List of parameters.

List of parameters.
}
\description{
Extracts model coefficients from objects with class
\code{\link[=GenFitMLVSBM]{GenFitMLVSBM}}

Extracts model coefficients from objects with class \code{\link[=FitMLVSBM]{FitMLVSBM}}
}
