% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMLVSBM-class.R
\docType{data}
\name{FitMLVSBM}
\alias{FitMLVSBM}
\title{An R6 Class ocject, a fitted multilevel network once $dovem() is done}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FitMLVSBM
}
\description{
An R6 Class ocject, a fitted multilevel network once $dovem() is done
}
\keyword{datasets}
