% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSBM-class.R
\docType{data}
\name{FitSBM}
\alias{FitSBM}
\title{An R6 Class ocject, a fitted level of a multilelvel network once $do_vem() is done}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FitSBM
}
\description{
An R6 Class ocject, a fitted level of a multilelvel network once $do_vem() is done
}
\keyword{datasets}
