% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenMLVSBM-class.R
\name{GenMLVSBM}
\alias{GenMLVSBM}
\title{R6Class for multilevel object}
\description{
Store all simulation parameters and list of fittedmodels.
Methods for global inference and model selection are included.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{nb_nodes}}{List of the umber of nodes for each levels}

\item{\code{simulation_parameters}}{List of parameters of the MLVSBM}

\item{\code{affiliation_matrix}}{Access the affiliation matrix}

\item{\code{adjacency_matrix}}{Access the list of adjacency_matrix}

\item{\code{memberships}}{Access the list of the clusterings}

\item{\code{fittedmodels}}{Get the list of selected fitted FitMLVSBM objects}

\item{\code{ICL}}{A summary table of selected fitted models and ICL model selection criterion}

\item{\code{ICL_sbm}}{Summary table of ICL by levels}

\item{\code{tmp_fittedmodels}}{A list of all fitted FitMLVSBM objects}

\item{\code{fittedmodels_sbm}}{A list of selected fitted FitSBM objects of each levels}

\item{\code{max_clusters}}{Access the list of maximum model size}

\item{\code{min_clusters}}{Access the list of minimum model size}

\item{\code{directed}}{Access the list of boolean for levels  direction}

\item{\code{directed}}{Access the list of the distribution used for each levels}

\item{\code{nb_levels}}{Access the number of levels in the network}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenMLVSBM-estimate_level}{\code{GenMLVSBM$estimate_level()}}
\item \href{#method-GenMLVSBM-estimate_sbm_neighbours}{\code{GenMLVSBM$estimate_sbm_neighbours()}}
\item \href{#method-GenMLVSBM-estimate_sbm_from_neighbours}{\code{GenMLVSBM$estimate_sbm_from_neighbours()}}
\item \href{#method-GenMLVSBM-estimate_sbm}{\code{GenMLVSBM$estimate_sbm()}}
\item \href{#method-GenMLVSBM-mcestimate}{\code{GenMLVSBM$mcestimate()}}
\item \href{#method-GenMLVSBM-estimate_neighbours}{\code{GenMLVSBM$estimate_neighbours()}}
\item \href{#method-GenMLVSBM-estimate_one}{\code{GenMLVSBM$estimate_one()}}
\item \href{#method-GenMLVSBM-estimate_all_bm}{\code{GenMLVSBM$estimate_all_bm()}}
\item \href{#method-GenMLVSBM-new}{\code{GenMLVSBM$new()}}
\item \href{#method-GenMLVSBM-findmodel}{\code{GenMLVSBM$findmodel()}}
\item \href{#method-GenMLVSBM-clearmodels}{\code{GenMLVSBM$clearmodels()}}
\item \href{#method-GenMLVSBM-addmodel}{\code{GenMLVSBM$addmodel()}}
\item \href{#method-GenMLVSBM-simulate}{\code{GenMLVSBM$simulate()}}
\item \href{#method-GenMLVSBM-clone}{\code{GenMLVSBM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-estimate_level"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-estimate_level}{}}}
\subsection{Method \code{estimate_level()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$estimate_level(
  level = NULL,
  Q_min = 1,
  Q_max = 10,
  Z = NULL,
  init = "hierarchical",
  depth = 1,
  nb_cores = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-estimate_sbm_neighbours"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-estimate_sbm_neighbours}{}}}
\subsection{Method \code{estimate_sbm_neighbours()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$estimate_sbm_neighbours(
  level = NULL,
  Q = NULL,
  Q_min = 1,
  Q_max = 10,
  fit = NULL,
  nb_cores = NULL,
  init = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-estimate_sbm_from_neighbours"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-estimate_sbm_from_neighbours}{}}}
\subsection{Method \code{estimate_sbm_from_neighbours()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$estimate_sbm_from_neighbours(
  level = NULL,
  Q = NULL,
  fits = NULL,
  nb_cores = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-estimate_sbm"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-estimate_sbm}{}}}
\subsection{Method \code{estimate_sbm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$estimate_sbm(level = NULL, Q = Q, Z = NULL, init = "hierarchical")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-mcestimate"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-mcestimate}{}}}
\subsection{Method \code{mcestimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$mcestimate(Q, Z = NULL, init = "hierarchical", independent = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-estimate_neighbours"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-estimate_neighbours}{}}}
\subsection{Method \code{estimate_neighbours()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$estimate_neighbours(
  level,
  fit = NULL,
  Q,
  independent = independent,
  nb_cores = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-estimate_one"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-estimate_one}{}}}
\subsection{Method \code{estimate_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$estimate_one(
  Q,
  Z = NULL,
  independent = FALSE,
  init = "hierarchical",
  nb_cores = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-estimate_all_bm"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-estimate_all_bm}{}}}
\subsection{Method \code{estimate_all_bm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$estimate_all_bm(
  Q = NULL,
  Z = NULL,
  independent = FALSE,
  clear = TRUE,
  nb_cores = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-new"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-new}{}}}
\subsection{Method \code{new()}}{
Constructor for R6 class MLVSBM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$new(
  n = NULL,
  X = NULL,
  A = NULL,
  L = NULL,
  Z = NULL,
  directed = NULL,
  sim_param = NULL,
  distribution = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{A list of size 2, the number of nodes}

\item{\code{X}}{A list of L adjacency matrices}

\item{\code{A}}{A list of L-1 affiliation matrices}

\item{\code{Z}}{A list of L vectors, the blocks membership}

\item{\code{directed}}{A vector of L booleans}

\item{\code{sim_param}}{A list of MLVSBM parameters for simulating networks}

\item{\code{distribution}}{The distributions of the interactions ("bernoulli")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A MLVSBM object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-findmodel"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-findmodel}{}}}
\subsection{Method \code{findmodel()}}{
Find a fitted model of a given size
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$findmodel(nb_clusters = NA, fit = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nb_clusters}}{A list of the size of the model}

\item{\code{fit}}{if fit = "best" return the best model according to the ICL}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A FitMLVSBM object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-clearmodels"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-clearmodels}{}}}
\subsection{Method \code{clearmodels()}}{
delete all fitted models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$clearmodels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-addmodel"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-addmodel}{}}}
\subsection{Method \code{addmodel()}}{
Added a FitMLVSBM object to the list of fitted model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$addmodel(fit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit}}{The FitMLVSBM object to be added}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-simulate}{}}}
\subsection{Method \code{simulate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$simulate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenMLVSBM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenMLVSBM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenMLVSBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
