% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{hierarClust}
\alias{hierarClust}
\title{Perform a Hierarchical Clustering}
\usage{
hierarClust(X, K)
}
\arguments{
\item{X}{An Adjacency Matrix}

\item{K}{the number of wanted clusters}
}
\value{
A vector : The clusters labels
}
\description{
Perform a Hierarchical Clustering
}
