% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultility.R, R/utility.R
\name{merge_clust}
\alias{merge_clust}
\title{Merge a list of clusters}
\usage{
merge_clust(Z, Q)

merge_clust(Z, Q)
}
\arguments{
\item{Z}{a vector of cluster memberships}

\item{Q}{the number of original clusters}
}
\value{
A list of Q(Q-1)/2 clustering of Q-1 clusters

A list of Q(Q-1)/2 clustering of Q-1 clusters
}
\description{
Merge a list of clusters

Merge a list of clusters
}
