% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_function.R
\name{mlvsbm_estimate_network}
\alias{mlvsbm_estimate_network}
\title{Infer a multilevel network (MLVSBM object), the original object is modified}
\usage{
mlvsbm_estimate_network(mlv)
}
\arguments{
\item{mlv}{A MLVSBM object, the network to be inferred}
}
\value{
A FitMLVSBM object, the best inference of the network
}
\description{
Infer a multilevel network (MLVSBM object), the original object is modified
}
