% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_function.R
\name{mlvsbm_simulate_network}
\alias{mlvsbm_simulate_network}
\title{Create a simulated multilevel network (MLVSBM object)}
\usage{
mlvsbm_simulate_network(n_I, n_O, Q_I, Q_O, pi_O, gamma, alpha_I, alpha_O,
  directed_I, directed_O, affiliation = "uniform",
  no_empty_org = FALSE)
}
\arguments{
\item{n_I}{A positive integer, the number of individuals}

\item{n_O}{A positive integer, the number of organisations}

\item{Q_I}{A positive integer, the number of clusters of individuals}

\item{Q_O}{A positive intger, the number of clusters of organisations}

\item{pi_O}{A probability vetor of length Q_O,
the mixture parameter for the organisations}

\item{gamma}{A Q_IxQ_O matrix with each column suming to one,
the mixture parameters for the individuals}

\item{alpha_I}{A Q_IxQ_I matrix giving the connectivity probabilities
of the individuals}

\item{alpha_O}{A Q_OxQ_O matrix giving the connectivity probabilities
of the organisations}

\item{directed_I}{A boolean, is level I a directed network?}

\item{directed_O}{A boolean, is level O a directed network?}

\item{affiliation}{The distribution under which the affiliation matrix is
simulated in c("uniform", "preferential")}

\item{no_empty_org}{A boolean with FALSE as default, should
every organisation have at least one affiliated individuals?
Needs to have n_I >= n_O}
}
\value{
An MLVSBM object, a simulated multilevel network with levels,
affiliations and memberships
}
\description{
Create a simulated multilevel network (MLVSBM object)
}
