% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGenMLVSBM-class.R
\name{plot.GenFitMLVSBM}
\alias{plot.GenFitMLVSBM}
\title{Generalized Multilevel SBM Plot}
\usage{
\method{plot}{GenFitMLVSBM}(x, type = c("graphon"), ...)
}
\arguments{
\item{x}{an R6 object of class \code{\link[=GenFitMLVSBM]{GenFitMLVSBM}}}

\item{type}{A string for the type of plot, just "matrix" for now}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
a ggplot2 object
}
\description{
basic matrix plot method for a FitMLVSBM object
}
\details{
Basic matrix plot method for a FitMLVSBM object
}
