% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot a matrix representation of the network.}
\usage{
plot_matrix(net, level = NULL)
}
\arguments{
\item{net}{A fitted network. Either a FitSBM or a FitMLVSBM object}

\item{level}{In \code{c("lower", "upper")}. Only for FitMLVSBM object.
If \code{NULL} then both level will be plotted.}
}
\value{
A \code{ggplot2} plot. The matrix representation of one or each level.
}
\description{
This function does not plot the affiliation relationship, only
the levels. The modes are ordered by blocks and a color gradient represent
the probability of each entries.
}
