% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGenMLVSBM-class.R
\name{predict.FitGenMLVSBM}
\alias{predict.FitGenMLVSBM}
\title{Model Predictions}
\usage{
\method{predict}{FitGenMLVSBM}(object, ...)
}
\arguments{
\item{object}{an R6 object of class \code{\link[=FitGenMLVSBM]{FitGenMLVSBM}}}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A list with the following entries:
\describe{
  \item{dyads}{A list of matrix with the probability of each dyads}
  \item{nodes}{A list of vectors with the clustering of each nodes}
}
}
\description{
Make predictions of dyads or give the nodes memberships for all levels
with a generalized multilevel SBM.
}
