% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGenMLVSBM-class.R, R/FitMLVSBM-class.R
\name{predict.FitMLVSBM}
\alias{predict.FitMLVSBM}
\title{Model Predictions}
\usage{
\method{predict}{FitMLVSBM}(object, ...)

\method{predict}{FitMLVSBM}(object, ...)
}
\arguments{
\item{object}{an R6 object of class \code{\link[=FitMLVSBM]{FitMLVSBM}}}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A list with the following entries:
\describe{
  \item{dyads}{A list of matrix with the probability of each dyads}
  \item{nodes}{A list of vectors with the clustering of each nodes}
}

A list with the following entries:
\describe{
  \item{dyads}{A list of matrix with the probability of each dyads}
  \item{nodes}{A list of vectors with the clustering of each nodes}
}
}
\description{
Make predictions from an SBM.

Make predictions from an SBM.
}
