% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_affiliation}
\alias{simulate_affiliation}
\title{Simulate of matrix of affiliation}
\usage{
simulate_affiliation(n, m, affiliation = "uniform", no_empty_org = FALSE)
}
\arguments{
\item{n}{An integer, the number of individuals}

\item{m}{An integer, the number of organizations}

\item{affiliation}{The type of affiliation between c("uniform", "preferential")}

\item{no_empty_org}{A Boolean. Force all columns to have at least a 1.
Must have \eqn{n>m}.}
}
\value{
A \eqn{n \times m} affiliation matrix, with a unique 1 on each rows
}
\description{
Simulate of matrix of affiliation
}
