% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultility.R
\name{spcClust}
\alias{spcClust}
\title{Perform a spectral clustering}
\usage{
spcClust(X, K)
}
\arguments{
\item{X}{an Adjacency matrix}

\item{K}{the number of clusters}
}
\value{
A vector : The clusters labels
}
\description{
Perform a spectral clustering
}
