% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{split_clust}
\alias{split_clust}
\title{Merge a list of clusters}
\usage{
split_clust(X, Z, Q)
}
\arguments{
\item{X}{an adjacency matrix}

\item{Z}{a vector of cluster memberships}

\item{Q}{The number of maximal clusters}
}
\value{
A list of Q clustering of Q+1 clusters
}
\description{
Merge a list of clusters
}
