% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMLVSBM-class.R
\name{FitMLVSBM}
\alias{FitMLVSBM}
\title{An R6 Class ocject, a fitted multilevel network once $dovem() is done}
\description{
An R6 Class ocject, a fitted multilevel network once $dovem() is done

An R6 Class ocject, a fitted multilevel network once $dovem() is done
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FitMLVSBM$new()}}
\item \href{#method-update_alpha}{\code{FitMLVSBM$update_alpha()}}
\item \href{#method-update_pi}{\code{FitMLVSBM$update_pi()}}
\item \href{#method-update_gamma}{\code{FitMLVSBM$update_gamma()}}
\item \href{#method-init_clustering}{\code{FitMLVSBM$init_clustering()}}
\item \href{#method-clear}{\code{FitMLVSBM$clear()}}
\item \href{#method-m_step}{\code{FitMLVSBM$m_step()}}
\item \href{#method-ve_step}{\code{FitMLVSBM$ve_step()}}
\item \href{#method-do_vem}{\code{FitMLVSBM$do_vem()}}
\item \href{#method-permute_empty_class}{\code{FitMLVSBM$permute_empty_class()}}
\item \href{#method-clone}{\code{FitMLVSBM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitMLVSBM$new(
  Q = list(I = 1, O = 1),
  A = NA,
  X = NA,
  M = list(I = NA, O = NA),
  directed = NA,
  distribution = list("bernoulli", "bernoulli"),
  independent = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_alpha"></a>}}
\if{latex}{\out{\hypertarget{method-update_alpha}{}}}
\subsection{Method \code{update_alpha()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitMLVSBM$update_alpha(safeguard = 2 * .Machine$double.eps)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_pi"></a>}}
\if{latex}{\out{\hypertarget{method-update_pi}{}}}
\subsection{Method \code{update_pi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitMLVSBM$update_pi(safeguard = 0.01)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_gamma"></a>}}
\if{latex}{\out{\hypertarget{method-update_gamma}{}}}
\subsection{Method \code{update_gamma()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitMLVSBM$update_gamma(safeguard = 1e-06)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init_clustering"></a>}}
\if{latex}{\out{\hypertarget{method-init_clustering}{}}}
\subsection{Method \code{init_clustering()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitMLVSBM$init_clustering(
  safeguard = 2 * .Machine$double.eps,
  method = "hierarchical",
  Z = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitMLVSBM$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-m_step"></a>}}
\if{latex}{\out{\hypertarget{method-m_step}{}}}
\subsection{Method \code{m_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitMLVSBM$m_step(safeguard = 1e-06)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ve_step"></a>}}
\if{latex}{\out{\hypertarget{method-ve_step}{}}}
\subsection{Method \code{ve_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitMLVSBM$ve_step(threshold = 1e-06, fixPointIter = 100, safeguard = 1e-06)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do_vem"></a>}}
\if{latex}{\out{\hypertarget{method-do_vem}{}}}
\subsection{Method \code{do_vem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitMLVSBM$do_vem(
  init = "hierarchical",
  threshold = 1e-06,
  maxIter = 1000,
  fixPointIter = 100,
  safeguard = 1e-06,
  Z = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-permute_empty_class"></a>}}
\if{latex}{\out{\hypertarget{method-permute_empty_class}{}}}
\subsection{Method \code{permute_empty_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitMLVSBM$permute_empty_class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitMLVSBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
