% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitSBM-class.R
\name{FitSBM}
\alias{FitSBM}
\title{An R6 Class ocject, a fitted level of a multilelvel network once $do_vem() is done}
\description{
An R6 Class ocject, a fitted level of a multilelvel network once $do_vem() is done

An R6 Class ocject, a fitted level of a multilelvel network once $do_vem() is done
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FitSBM$new()}}
\item \href{#method-update_alpha}{\code{FitSBM$update_alpha()}}
\item \href{#method-update_pi}{\code{FitSBM$update_pi()}}
\item \href{#method-init_clustering}{\code{FitSBM$init_clustering()}}
\item \href{#method-m_step}{\code{FitSBM$m_step()}}
\item \href{#method-ve_step}{\code{FitSBM$ve_step()}}
\item \href{#method-do_vem}{\code{FitSBM$do_vem()}}
\item \href{#method-permute_empty_class}{\code{FitSBM$permute_empty_class()}}
\item \href{#method-clear}{\code{FitSBM$clear()}}
\item \href{#method-clone}{\code{FitSBM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitSBM$new(
  Q = 1,
  X = NULL,
  M = NULL,
  directed = FALSE,
  distribution = "bernoulli"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_alpha"></a>}}
\if{latex}{\out{\hypertarget{method-update_alpha}{}}}
\subsection{Method \code{update_alpha()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitSBM$update_alpha(safeguard = 1e-06)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_pi"></a>}}
\if{latex}{\out{\hypertarget{method-update_pi}{}}}
\subsection{Method \code{update_pi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitSBM$update_pi(safeguard = 1e-06)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init_clustering"></a>}}
\if{latex}{\out{\hypertarget{method-init_clustering}{}}}
\subsection{Method \code{init_clustering()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitSBM$init_clustering(safeguard = 1e-06, method = "hierarchical", Z = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-m_step"></a>}}
\if{latex}{\out{\hypertarget{method-m_step}{}}}
\subsection{Method \code{m_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitSBM$m_step(safeguard = 1e-06)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ve_step"></a>}}
\if{latex}{\out{\hypertarget{method-ve_step}{}}}
\subsection{Method \code{ve_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitSBM$ve_step(threshold = 1e-06, fixPointIter = 100, safeguard = 1e-06)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do_vem"></a>}}
\if{latex}{\out{\hypertarget{method-do_vem}{}}}
\subsection{Method \code{do_vem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitSBM$do_vem(
  init = "hierarchical",
  threshold = 1e-06,
  maxIter = 1000,
  fixPointIter = 100,
  safeguard = 1e-06,
  Z = NULL,
  bound = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-permute_empty_class"></a>}}
\if{latex}{\out{\hypertarget{method-permute_empty_class}{}}}
\subsection{Method \code{permute_empty_class()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitSBM$permute_empty_class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitSBM$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitSBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
