% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLVSBM-class.R
\docType{data}
\name{MLVSBM}
\alias{MLVSBM}
\title{An R6Class for multilevel object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MLVSBM
}
\description{
An R6Class for multilevel object
}
\keyword{datasets}
