% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalize-userfunction.R
\name{mlvsbm_create_generalized_network}
\alias{mlvsbm_create_generalized_network}
\title{Create a GenMLVSBM object from observed data}
\usage{
mlvsbm_create_generalized_network(X, A, directed = NULL, distribution = NULL)
}
\arguments{
\item{X}{A list of \eqn{L} square matrices with binary or count data,
ordered from upper to lower level}

\item{A}{A list of \eqn{L-1} matrices the affiliation matrix, linking X[[l]]
and X[[l+1]] with \code{nrow(A[[l]]) == nrow(X[[l+1]])} and
\code{ncol(A[[l]]) == nrow(X[[l]])}. Rows may have any number of affiliations
but will be rescaled so that each rows sum either to 1 or 0.}

\item{directed}{A vector of L booleans are the levels
directed or not. Default will check if the matrix are symmetric or not.}

\item{distribution}{A vector of length L, the distribution of X,
only "bernoulli" and "poisson" are implemented.}
}
\value{
An unfitted MLVSBM object corresponding to the multilevel network
}
\description{
Create a GenMLVSBM object from observed data
}
\examples{
ind_adj <- matrix(stats::rbinom(n = 10**2, size = 1, prob = .2),
                  nrow = 10, ncol = 10)
org_adj <- matrix(stats::rbinom(n = 10**2, size = 1, prob = .3),
                  nrow = 10, ncol = 10)
affiliation <- diag(1, 10)
my_mlvsbm <- mlvsbm_create_generalized_network(X = list(org_adj, ind_adj),
                                   directed = list(I = FALSE, O = FALSE),
                                   A = list(affiliation),
                                   distribution = rep("bernoulli", 2))
}
