% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_function.R
\name{mlvsbm_create_network}
\alias{mlvsbm_create_network}
\title{Create a MLVSBM object from observed data}
\usage{
mlvsbm_create_network(X_I, X_O, A)
}
\arguments{
\item{X_I}{A matrix square binary adjacency, the individual level}

\item{X_O}{A matrix square binary adjacency, the organisational level}

\item{A}{A matrix the affiliation matrix with individuals in rows and
organisations in columns}
}
\value{
An unfitted MLVSBM object corresponding to the multilevel network
}
\description{
Create a MLVSBM object from observed data
}
