% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_function.R
\name{mlvsbm_simulate_network}
\alias{mlvsbm_simulate_network}
\title{Create a simulated multilevel network (MLVSBM object)}
\usage{
mlvsbm_simulate_network(
  n,
  Q,
  pi,
  gamma,
  alpha,
  directed,
  affiliation = "uniform",
  distribution = list("bernoulli", "bernoulli"),
  no_empty_org = FALSE
)
}
\arguments{
\item{n}{A list of 2 positive integers,
the number of individuals and organisations}

\item{Q}{A list of 2 positive integers,
the number of clusters of individuals and organisations}

\item{pi}{A vector of probabilities of length Q_O,
the mixture parameter for the organisations}

\item{gamma}{A Q_IxQ_O matrix with each column suming to one,
the mixture parameters for the individuals}

\item{alpha}{A list of 2 matrices, a Q_IxQ_I matrix giving the connectivity probabilities
of the individuals and a Q_OxQ_O matrix giving the connectivity probabilities
of the organisations}

\item{directed}{A list of 2 booleans. Is the individual level a directed network ?
Is the organisational level a directed network ?}

\item{affiliation}{The distribution under which the affiliation matrix is
simulated in c("uniform", "preferential")}

\item{distribution}{list for the distribution of X,
only "bernoulli" is implemented}

\item{no_empty_org}{A boolean with FALSE as default, should
every organisation have at least one affiliated individuals?
Needs to have n_I >= n_O}
}
\value{
An MLVSBM object, a simulated multilevel network with levels,
affiliations and memberships
}
\description{
Create a simulated multilevel network (MLVSBM object)
}
