% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMLVSBM-class.R
\name{plot.FitMLVSBM}
\alias{plot.FitMLVSBM}
\title{Multilevel SBM Plot}
\usage{
\method{plot}{FitMLVSBM}(x, type = c("matrix"), ...)
}
\arguments{
\item{x}{an R6 object of class \code{\link[=FitMLVSBM]{FitMLVSBM}}}

\item{type}{A string for the type of plot, just "matrix" for now}

\item{...}{additional parameters. block ordering with
order = c("affiliation", "degree", "natural")}
}
\value{
a ggplot2 object
}
\description{
basic matrix plot method for a FitMLVSBM object
}
\details{
Basic matrix plot method for a FitMLVSBM object
}
