% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_adjacency}
\alias{simulate_adjacency}
\title{Simulation an adjacency matrix}
\usage{
simulate_adjacency(
  Z,
  n,
  alpha,
  directed,
  distribution = "bernoulli",
  no_isolated_node = FALSE
)
}
\arguments{
\item{Z}{A vector of integer of size n, the label}

\item{n}{An integer, the number of rows or columbns of the matrix}

\item{alpha}{A max(Z)xmax(Z) matrix, the connectivity parameters}

\item{directed}{A boolean, Is the network directed or not ?}

\item{distribution}{The distribution of the indices: only "bernoulli"}

\item{no_isolated_node}{A boolean, may row and column of adjacency matrices sum to 0}
}
\value{
A nxn adjacency matrix
}
\description{
Simulation an adjacency matrix
}
